Replace checksum with last modification time in grid file
The filename for grid files no longer uses the checksum of the orders table. Instead, it now uses the last modification time of the orders. This change also includes a predicate to ensure that the file only gets outputted if it is up-to-date with the last changes in the orders.