Refactor order step management to Steps class
The functionality to get and set the current step in the order process has been moved from the Orders class to the Steps class. This is a structural change aimed at better organizing related pieces of code. Calls to these functions in the codebase have been updated to use the new Steps class.