Refactor login check in make_selection.php
The mds_wp_login_check function has been refactored to use WordPress' native function for checking if a user is logged in. Now, if a user is not logged in and attempts to access a page that requires authentication, a JSON response with an error and redirect URL will be returned, offering a more informative and user-friendly experience.