Refactor user login validation in Forms class
The logic in the 'mds_form_submission' method of the Forms class was refactored to not only check for the existence of 'mds_dest' POST data, but also validate if the user is logged in. This change enhances the security by preventing unauthorized users from attempting to submit a form.