Update plugin version to 2.5.10.1.
The version of the plugin was updated to 2.5.10.1. This update includes removing the old 2.5.10 version documentation from all files. The plugin URI, stable tag, and MDS_VERSION constants have been updated with the new version number. This adds another digit to the end of the version that will be used for automatically incrementing the dev branch commits in an attempt to allow dev branch updates.