Improve time calculation for order expiration
The old version was showing the remaining time wrong until an order would be canceled due to not being renewed. This update revises the process to compute the remaining time in days, hours, and minutes and display this breakdown to the user. Additionally, suited the SQL query to WordPress standards and applied appropriate escaping functions to improve security.